@@ -1,6 +1,6 @@
 {
   "name": "mongodb",
-  "version": "6.10.0",
+  "version": "6.11.0",
   "description": "The official MongoDB driver for Node.js",
   "main": "lib/index.js",
   "files": [
@@ -25,8 +25,8 @@
     "email": "dbx-node@mongodb.com"
   },
   "dependencies": {
-    "@mongodb-js/saslprep": "^1.1.5",
-    "bson": "^6.7.0",
+    "@mongodb-js/saslprep": "^1.1.9",
+    "bson": "^6.10.0",
     "mongodb-connection-string-url": "^3.0.0"
   },
   "peerDependencies": {
@@ -62,42 +62,42 @@
     }
   },
   "devDependencies": {
-    "@aws-sdk/credential-providers": "^3.632.0",
+    "@aws-sdk/credential-providers": "^3.686.0",
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.47.5",
+    "@microsoft/api-extractor": "^7.47.11",
     "@microsoft/tsdoc-config": "^0.17.0",
-    "@mongodb-js/zstd": "^1.2.0",
+    "@mongodb-js/zstd": "^1.2.2",
     "@types/chai": "^4.3.17",
     "@types/chai-subset": "^1.3.5",
     "@types/express": "^4.17.21",
     "@types/kerberos": "^1.1.5",
-    "@types/mocha": "^10.0.7",
-    "@types/node": "^22.4.1",
+    "@types/mocha": "^10.0.9",
+    "@types/node": "^22.9.0",
     "@types/saslprep": "^1.0.3",
     "@types/semver": "^7.5.8",
     "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^3.2.12",
     "@types/whatwg-url": "^11.0.5",
-    "@typescript-eslint/eslint-plugin": "^8.2.0",
-    "@typescript-eslint/parser": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "8.4.0",
+    "@typescript-eslint/parser": "8.4.0",
     "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^9.0.0",
+    "eslint": "9.9.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-mocha": "^10.4.1",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-tsdoc": "^0.3.0",
-    "eslint-plugin-unused-imports": "^4.0.0",
-    "express": "^4.19.2",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "express": "^4.21.1",
     "gcp-metadata": "^5.3.0",
     "js-yaml": "^4.1.0",
-    "mocha": "^10.4.0",
+    "mocha": "^10.8.2",
     "mocha-sinon": "^2.1.2",
     "mongodb-client-encryption": "^6.1.0",
-    "mongodb-legacy": "^6.1.0",
+    "mongodb-legacy": "^6.1.3",
     "nyc": "^15.1.0",
     "prettier": "^3.3.3",
     "semver": "^7.6.3",
@@ -107,7 +107,7 @@
     "socks": "^2.8.1",
     "source-map-support": "^0.5.21",
     "ts-node": "^10.9.2",
-    "tsd": "^0.31.1",
+    "tsd": "^0.31.2",
     "typescript": "5.5",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.3.1",
@@ -151,6 +151,7 @@
     "check:oidc-test": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc.prose.test.ts",
     "check:oidc-azure": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_azure.prose.05.test.ts",
     "check:oidc-gcp": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_gcp.prose.06.test.ts",
+    "check:oidc-k8s": "mocha --config test/mocha_mongodb.json test/integration/auth/mongodb_oidc_k8s.prose.07.test.ts",
     "check:ocsp": "mocha --config test/manual/mocharc.json test/manual/ocsp_support.test.js",
     "check:kerberos": "nyc mocha --config test/manual/mocharc.json test/manual/kerberos.test.ts",
     "check:tls": "mocha --config test/manual/mocharc.json test/manual/tls_support.test.ts",
