@@ -521,4 +521,4 @@ router.post('/getNotification', auth, async (req, res) => {
 });
 
 
-module.exports = router;
+export default router;
