@@ -1,14 +1,13 @@
-import express from 'express';
-import cors from 'cors';
-import mongoose from 'mongoose';
-import { join } from 'path';
-import dotenv from 'dotenv';
-import userRoutes from './routes/routes.js';
-import forgotPasswordRoutes from './routes/forgotPassword.js';
+const express = require('express');
+const cors = require('cors');
+const mongoose = require('mongoose');
+const join = require('path');
+const dotenv = require('dotenv');
+const userRoutes = ('./routes/routes.js');
+const forgotPasswordRoutes = require('./routes/forgotPassword.js');
 
 const app = express();
 const PORT = process.env.PORT || 5000;
-
 
 // Middleware
 app.use(express.json());
@@ -18,10 +17,7 @@ dotenv.config();
 app.use('/api', userRoutes);
 app.use('/api/forgot-password', forgotPasswordRoutes);
 
-mongoose.connect(process.env.MONGO_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-})
+mongoose.connect(process.env.MONGO_URI)
   .then(() => console.log('Connected to MongoDB'))
   .catch((err) => console.log('Failed to connect to MongoDB:', err));
 
